+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 
+- The number of threads used by FlowMachine to manage running queries is no longer the same as the number of database connections. It can instead be controlled using the `THREAD_POOL_SIZE` environment variable, and will default to `5*n_cores`.
+
 ### Fixed
 
 ### Removed